% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_business_days.R
\name{shptogf}
\alias{shptogf}
\title{prepare for ggplot shape file to ggplot style}
\usage{
shptogf(shp)
}
\arguments{
\item{shp}{spatial polygon data frame}
}
\description{
prepare for ggplot shape file to ggplot style
}
